owner order creator initialized using past components, working on recator for desired functionality
