+++ front-end/tons-of-tacos/components/owner-dashboard/action-bar/action-bar.module.css
@@ -1,0 +1,9 @@
+.actionBar {
+  display: flex;
+}
+
+.ownerCreate {
+  margin-left: 3rem;
+  font-size: 1rem;
+  font-weight: bold;
+}
