+++ front-end/tons-of-tacos/components/ui/cards/card.module.css
@@ -22,8 +22,9 @@
   background-color: var(--background-color);
   border-radius: 10px;
   width: 99%;
-  height: 99.5%;
+  height: 100%;
   justify-self: center;
+  /* padding: 1rem; */
 }
 
 .expand {
